Split CustomThemeProvider
